+++ src/main/java/com/example/doctorcare/api/service/TimeDoctorService.java
@@ -21,7 +21,7 @@ public class TimeDoctorService {
     public void save(TimeDoctors timeDoctors){
         timeDoctorRepository.save(timeDoctorsMapper.convertToEntity(timeDoctors));
     }
-    public List<TimeDoctors> findAllByDoctor(int doctorId){
+    public List<TimeDoctors> findAllByDoctor(Long doctorId){
         List<TimeDoctors> timeDoctorsList= new ArrayList<>();
         timeDoctorRepository.findByDoctor_Id(doctorId).forEach(timeDoctors -> {timeDoctorsList.add(timeDoctorsMapper.convertToDto(timeDoctors));});
     return timeDoctorsList;
